requirements.txt
absl-py==0.2.2
astor==0.7.1
bleach==1.5.0
certifi==2016.2.28
cycler==0.10.0
enum34==1.1.6
gast==0.2.0
grpcio==1.13.0
h5py==2.8.0
html5lib==0.9999999
kiwisolver==1.0.1
mahotas==1.4.4
Markdown==2.6.11
matplotlib==2.2.3
numpy==1.15.1
opencv-contrib-python==3.4.2.17
opencv-python==3.4.2.17
pandas==0.23.3
Pillow==5.1.0
protobuf==3.6.0
py-glcm==1.1a0
pyparsing==2.2.0
python-dateutil==2.7.3
pytz==2018.5
scikit-learn==0.19.1
scipy==1.1.0
six==1.11.0
tensorboard==1.8.0
tensorflow==1.8.0
tensorflow-gpu==1.8.0
tensorflow-tensorboard==0.4.0
termcolor==1.1.0
torch==0.4.1
torchvision==0.2.1
Werkzeug==0.14.1
